requirements.txt
python-dotenv==1.0.1
redis==5.2.1
ShopifyAPI==12.7.0
requests==2.32.3